Sync r4739 trunk to new svn
git-svn-id: https://svn.apache.org/repos/asf/incubator/openmeetings/trunk/singlewebapp@1220406 13f79535-47bb-0310-9956-ffa450edef68

+++ src/app/org/openmeetings/app/data/flvrecord/converter/FlvInterviewConverter.java
@@ -388,7 +388,7 @@ public class FlvInterviewConverter extends BaseConverter {
 				currentTimeInMilliSeconds += 1000;
 
 				double cLength = 100 * ((double) currentTimeInMilliSeconds)
-						/ ((double) completeLengthInSeconds);
+						/ completeLengthInSeconds;
 
 				int progress = Double.valueOf(cLength).intValue();
